requirements.txt
numpy==2.4.2
scipy==1.17.1
matplotlib==3.10.8
seaborn==0.13.2
pyyaml==6.0.3
flask==3.1.3
flask-socketio==5.6.1
